Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Here are my results.  Since I have no bluetooth headset or a2dp devices I tested in Windows 7 with a bluetooth dongle.  With the ifdefs in place the audio had a lot of breakage with a2dp.  With the ifdefs removed the audio quality was much improved.

Patch-set: 1