Update patch set 1
Patch Set 1:

I don't care if the BT address incorrectly reports BTADDR of "monkeyspunk" if "monkeyspunk" was used as a hack to get things working properly. This condition is inherited from board-mahimahi.c (passion) and there may or may not be a reason for this, so unless this change presents a positive functional impact, I don't want to merge it. What functional problem does this address?

Was this tested on a variety of BT devices to verify correct operation, and if so, what are the functional benefits?

Patch-set: 1